Move estimation related code to estimate.py
