+++ causing/estimate.py
@@ -4,11 +4,14 @@
 # pylint: disable=invalid-name
 # pylint: disable=len-as-condition
 
+from copy import copy, deepcopy
+
 import numdifftools as nd
 import numpy as np
-from copy import deepcopy
-from numpy import allclose, array_equal, diag, eye, linspace, zeros
+from numpy import allclose, array_equal, diag, eye, linspace, zeros, isnan
 from numpy.linalg import cholesky, inv, LinAlgError
+import torch
+from scipy.optimize import minimize
 
 from causing import utils
 
@@ -208,7 +211,7 @@ def compute_cov_direct(sse_hat, hessian_hat, model_dat):
 def check_estimate_effects(model_dat, do_print=True):
     """estimate structural model given alpha in model_dat"""
 
-    mx_hat, my_hat, sse_hat = utils.estimate_snn(model_dat, do_print)
+    mx_hat, my_hat, sse_hat = estimate_snn(model_dat, do_print)
 
     ex_hat, ey_hat = utils.total_effects_alg(
         mx_hat, my_hat, model_dat["edx"], model_dat["edy"]
@@ -438,7 +441,7 @@ def estimate_effects(model_dat):
         ey_hat,
     ) = check_estimate_effects(model_dat)
     # automatic Hessian
-    hessian = utils.sse_hess(mx_hat, my_hat, model_dat)
+    hessian = sse_hess(mx_hat, my_hat, model_dat)
     # numeric Hessian
     hessian_num = sse_hess_num(mx_hat, my_hat, model_dat)
 
@@ -462,14 +465,12 @@ def estimate_effects(model_dat):
     cov_direct_hat = compute_cov_direct(sse_hat, hessian_hat, model_dat)
 
     # compute estimated direct, total and mediation effects and standard deviations
-    mx_hat_std, my_hat_std = utils.compute_direct_std(cov_direct_hat, model_dat)
-    ex_hat_std, ey_hat_std = utils.total_effects_std(
-        direct_hat, cov_direct_hat, model_dat
-    )
+    mx_hat_std, my_hat_std = compute_direct_std(cov_direct_hat, model_dat)
+    ex_hat_std, ey_hat_std = total_effects_std(direct_hat, cov_direct_hat, model_dat)
     exj_hat, eyj_hat, eyx_hat, eyy_hat = utils.compute_mediation_effects(
         mx_hat, my_hat, ex_hat, ey_hat, model_dat["yvars"], model_dat["final_var"]
     )
-    (exj_hat_std, eyj_hat_std, eyx_hat_std, eyy_hat_std) = utils.compute_mediation_std(
+    (exj_hat_std, eyj_hat_std, eyx_hat_std, eyy_hat_std) = compute_mediation_std(
         ex_hat_std,
         ey_hat_std,
         eyx_hat,
@@ -511,7 +512,7 @@ def estimate_biases(model_dat):
     biases_std = zeros(model_dat["ndim"])
     for bias_ind in range(model_dat["ndim"]):
         # compute biases
-        bias, hess_i, sse = utils.optimize_biases(model_dat, bias_ind)
+        bias, hess_i, sse = optimize_biases(model_dat, bias_ind)
         biases[bias_ind] = bias
 
         # compute biases_std
@@ -535,3 +536,350 @@ def estimate_models(model_dat):
         estimate_dat["biases_std"] = biases_std
 
     return estimate_dat
+
+
+class StructuralNN(torch.nn.Module):
+    """AD identified structural linear nn,
+    linear ychat approximation using ex effects reduced form"""
+
+    def __init__(self, model_dat):
+        super(StructuralNN, self).__init__()
+
+        self.eye = torch.DoubleTensor(eye(model_dat["ndim"]))
+        self.idx = torch.DoubleTensor(model_dat["idx"])
+        self.idy = torch.DoubleTensor(model_dat["idy"])
+        self.xcdat = torch.DoubleTensor(model_dat["xcdat"])
+
+    def forward(self, mx, my):
+        # impose identification restrictions already on imput
+        # ToDo: use torch.nn.utils.prune custom_from_mask or own custom method
+        mx = mx * self.idx
+        my = my * self.idy
+
+        ey = (self.eye - my).inverse()
+        ex = ey @ mx
+        dy = ex @ self.xcdat  # reduced form
+        ychat = dy
+
+        return ychat
+
+
+def optimize_ssn(
+    ad_model, mx, my, fym, ydata, selwei, model_dat, optimizer, params, do_print=True
+):
+    """ad torch optimization of structural neural network"""
+
+    # parameters
+    rel = 0.0001  # ToDo: define globally
+    nr_conv_min = 5  # ToDo: define globally
+
+    sse = torch.DoubleTensor([0])
+    sse_old = torch.DoubleTensor([1])
+    nr_conv = 0
+    epoch = 0
+    while nr_conv < nr_conv_min:
+        sse_old = copy(sse)
+        ychat = ad_model(*params)
+        sse = sse_orig(mx, my, fym, ychat, ydata, selwei, model_dat)  # forward
+        optimizer.zero_grad()
+        sse.backward(create_graph=True)  # backward
+        optimizer.step()
+        if abs(sse - sse_old) / sse_old < rel:
+            nr_conv += 1
+        else:
+            nr_conv = 0
+        nrm = sum([torch.norm(param) for param in params]).detach().numpy()  # type: ignore
+        if do_print:
+            print(
+                "epoch {:>4}, sse {:10f}, param norm {:10f}".format(
+                    epoch, sse.item(), nrm
+                )
+            )
+        epoch += 1
+
+    return sse
+
+
+def estimate_snn(model_dat, do_print=True):
+    """estimate direct effects in identified structural form
+    using PyTorch AD automatic differentiation
+
+    forcasting y is done by reduced form since it is already solved for dy
+    structural form:
+        dy = my @ dy + mx @ dx
+        mx, my is a linear network of at most ndim + mdim layers of max of max dims max(ndim, mdim)
+        with identifiying restrictions idx, idy
+    reduced form:
+        dy = ex @ dx
+        ex is a linear network with one layer of dimension (ndim, mdim)
+        with restrictions edx
+    Estimating effects with automatic differentiation only works for DAG
+    """
+
+    fym = torch.DoubleTensor(model_dat["fym"])
+    selwei = torch.DoubleTensor(model_dat["selwei"])
+
+    # start at theoretical direct effects
+    mx = torch.DoubleTensor(deepcopy(model_dat["mx_theo"]))
+    my = torch.DoubleTensor(deepcopy(model_dat["my_theo"]))
+
+    # define optimization parameters
+    ydata = torch.DoubleTensor(model_dat["ymcdat"])  # ymcdat
+    mx.requires_grad_(True)
+    my.requires_grad_(True)
+    params = [mx, my]
+    ad_model = StructuralNN(model_dat)  # ychat
+    # Adam, Adadelta, Adagrad, AdamW, Adamax, RMSprop, Rprop
+    optimizer = torch.optim.Rprop(params)
+
+    if do_print:
+        print(
+            "\nEstimation of direct effects using a structural neural network \n"
+            "with regularization parameter alpha = {:10f}:".format(model_dat["alpha"])
+        )
+    sse = optimize_ssn(
+        ad_model, mx, my, fym, ydata, selwei, model_dat, optimizer, params, do_print
+    )
+
+    mx = mx.detach().numpy()
+    my = my.detach().numpy()
+    sse = sse.detach().numpy()
+    assert allclose(
+        mx, mx * model_dat["idx"]
+    ), "idx identification restrictions not met:\n{}\n!=\n{}".format(
+        mx, mx * model_dat["idx"]
+    )
+    assert allclose(
+        my, my * model_dat["idy"]
+    ), "idy identification restrictions not met:\n{}\n!=\n{}".format(
+        my, my * model_dat["idy"]
+    )
+
+    return mx, my, sse
+
+
+def sse_bias(bias, bias_ind, model_dat):
+    """sum of squared errors given modification indicator, Tikhonov not used"""
+
+    yhat = model_dat["old_model"](model_dat["xdat"], bias, bias_ind)
+    ymhat = model_dat["fym"] @ yhat
+    err = ymhat - model_dat["ymdat"]
+    sse = np.sum(err * err * diag(model_dat["selwei"]).reshape(-1, 1))
+
+    print("sse {:10f}, bias {:10f}".format(sse, float(bias)))
+
+    return sse
+
+
+def optimize_biases(model_dat, bias_ind):
+    """numerical optimize modification indicator for single equation"""
+
+    # optimizations parameters
+    bias_start = 0
+    method = "SLSQP"  # BFGS, SLSQP, Nelder-Mead, Powell, TNC, COBYLA, CG
+
+    print("\nEstimation of bias for {}:".format(model_dat["yvars"][bias_ind]))
+    out = minimize(sse_bias, bias_start, args=(bias_ind, model_dat), method=method)
+
+    bias = out.x
+    sse = out.fun
+
+    if hasattr(out, "hess_inv"):
+        hess_i = inv(out.hess_inv)
+        print("Scalar Hessian from method {}.".format(method))
+    else:
+        hess_i = nd.Derivative(sse_bias, n=2)(bias, bias_ind, model_dat)
+        print("Scalar Hessian numerically.")
+
+    return bias, hess_i, sse
+
+
+def sse_hess(mx, my, model_dat):
+    """compute automatic Hessian of sse at given data and direct effects"""
+
+    fym = torch.DoubleTensor(model_dat["fym"])
+    ydata = torch.DoubleTensor(model_dat["ymcdat"])
+    selwei = torch.DoubleTensor(model_dat["selwei"])
+
+    def sse_orig_vec_alg(direct):
+        """computes the ad target function sum of squared errors,
+        input as tensor vectors, yields Hessian in usual dimension of identified parameters"""
+        mx, my = utils.directmat(direct, model_dat["idx"], model_dat["idy"])
+        ad_model = StructuralNN(model_dat)
+        ychat = ad_model(mx, my)
+        return sse_orig(mx, my, fym, ychat, ydata, selwei, model_dat)
+
+    direct = utils.directvec(mx, my, model_dat["idx"], model_dat["idy"])
+    hessian = torch.autograd.functional.hessian(sse_orig_vec_alg, direct)
+
+    # symmetrize Hessian, such that numerically well conditioned
+    hessian = hessian.detach().numpy()
+    hessian = (hessian + hessian.T) / 2
+
+    return hessian
+
+
+def tvals(eff, std):
+    """compute t-values by element wise division of eff and std matrices"""
+
+    assert eff.shape == std.shape
+
+    if len(eff.shape) == 1:  # vector
+        rows = eff.shape[0]
+        tvalues = np.empty(rows) * np.nan
+        for i in range(rows):
+            if std[i] != 0:
+                tvalues[i] = eff[i] / std[i]
+
+    if len(eff.shape) == 2:  # matrix
+        rows, cols = eff.shape
+        tvalues = zeros((rows, cols))
+        for i in range(rows):
+            for j in range(cols):
+                if std[i, j] != 0:
+                    tvalues[i, j] = eff[i, j] / std[i, j]
+                else:
+                    tvalues[i, j] = np.nan
+
+    return tvalues
+
+
+def compute_mediation_std(ex_hat_std, ey_hat_std, eyx, eyy, yvars, final_var):
+    """compute mediation std"""
+
+    # dimensions
+    ndim = ex_hat_std.shape[0]
+    mdim = ex_hat_std.shape[1]
+    jvar = list(yvars).index(final_var)
+
+    exj_hat_std = ex_hat_std[jvar, :]  # (mdim)
+    eyj_hat_std = ey_hat_std[jvar, :]  # (ndim)
+
+    # construct matices of repeating rows
+    exj_hat_std_mat = np.tile(exj_hat_std, (ndim, 1))  # (ndim x mdim)
+    eyj_hat_std_mat = np.tile(eyj_hat_std, (ndim, 1))  # (ndim x ndim)
+
+    # column sums of mediation matrices
+    x_colsum = np.sum(eyx, axis=0)
+    y_colsum = np.sum(eyy, axis=0)
+    # normed mediation matrices by division by column sums,
+    # zero sum for varaibles w/o effect on others
+    # substituted by nan to avoid false interpretation
+    x_colsum[x_colsum == 0] = np.nan
+    y_colsum[y_colsum == 0] = np.nan
+    eyx_colnorm = zeros((ndim, mdim))  # (ndim x mdim)
+    eyx_colnorm[:] = np.nan
+    for j in range(mdim):
+        if not isnan(x_colsum[j]):
+            eyx_colnorm[:, j] = eyx[:, j] / x_colsum[j]
+    eyy_colnorm = zeros((ndim, ndim))  # (ndim x ndim)
+    eyy_colnorm[:] = np.nan
+    for j in range(ndim):
+        if not isnan(y_colsum[j]):
+            eyy_colnorm[:, j] = eyy[:, j] / y_colsum[j]
+
+    # mediation std matrices
+    eyx_hat_std = exj_hat_std_mat * eyx_colnorm  # (ndim x mdim)
+    eyy_hat_std = eyj_hat_std_mat * eyy_colnorm  # (ndim x ndim)
+
+    return exj_hat_std, eyj_hat_std, eyx_hat_std, eyy_hat_std
+
+
+def compute_direct_std(vcm_direct_hat, model_dat):
+    """compute direct effects standard deviations from direct effects covariance matrix"""
+
+    direct_std = diag(vcm_direct_hat) ** (1 / 2)
+    mx_std, my_std = utils.directmat_alg(direct_std, model_dat["idx"], model_dat["idy"])
+
+    return mx_std, my_std
+
+
+def total_effects_std(direct_hat, vcm_direct_hat, model_dat):
+    """compute total effects standard deviations
+
+    given estimated vcm_direct_hat,
+    using algebraic delta method for covariance Matrix of effects and
+    algebraic gradient of total effects wrt. direct effects
+    """
+
+    # compute vec matrices for algebraic effects gradient wrt. to direct_hat
+    vecmaty = utils.vecmat(model_dat["idy"])
+    vecmatx = utils.vecmat(model_dat["idx"])
+    vecmaty = np.hstack(
+        (vecmaty, zeros((model_dat["ndim"] * model_dat["ndim"], model_dat["qxdim"])))
+    )
+    vecmatx = np.hstack(
+        (zeros((model_dat["ndim"] * model_dat["mdim"], model_dat["qydim"])), vecmatx)
+    )
+
+    # compute algebraic gradient of total effects wrt. direct effects
+    mx, my = utils.directmat_alg(direct_hat, model_dat["idx"], model_dat["idy"])
+    ey = inv(eye(model_dat["ndim"]) - my)
+    jac_effects_y = (
+        np.kron(ey.T, ey) - eye(model_dat["ndim"] * model_dat["ndim"])
+    ) @ vecmaty + vecmaty
+    jac_effects_x = (
+        np.kron((ey @ mx).T, ey) @ vecmaty
+        + np.kron(eye(model_dat["mdim"]), (ey - eye(model_dat["ndim"]))) @ vecmatx
+        + vecmatx
+    )
+
+    # reduce to rows corresponding to nonzero effects
+    indy = (
+        np.reshape(model_dat["edy"], model_dat["ndim"] * model_dat["ndim"], order="F")
+        != 0
+    )
+    indx = (
+        np.reshape(model_dat["edx"], model_dat["ndim"] * model_dat["mdim"], order="F")
+        != 0
+    )
+    jac_effects_y = jac_effects_y[indy, :]
+    jac_effects_x = jac_effects_x[indx, :]
+    jac_effects = np.vstack((jac_effects_y, jac_effects_x))
+
+    # compare numeric and algebraic gradient of effects
+    do_compare = True
+    if do_compare:
+        jac_effects_num = nd.Jacobian(utils.total_from_direct)(
+            direct_hat,
+            model_dat["idx"],
+            model_dat["idy"],
+            model_dat["edx"],
+            model_dat["edy"],
+        )
+        jac_effects_num = jac_effects_num.reshape(jac_effects.shape)
+        atol = 10 ** (-4)  # instead of default 10**(-8)
+        print(
+            "Numeric and algebraic gradient of total effects wrt. direct effects allclose: {}.".format(
+                allclose(jac_effects_num, jac_effects, atol=atol)
+            )
+        )
+
+    # algebraic delta method effects covariance Matrix
+    vcm_effects = jac_effects @ vcm_direct_hat @ jac_effects.T
+    effects_std = diag(vcm_effects) ** (1 / 2)
+    ex_std, ey_std = utils.directmat_alg(
+        effects_std, model_dat["edx"], model_dat["edy"]
+    )
+    # set main diag of ey_std to 0, since edy diag is 1 instead of 0
+    np.fill_diagonal(ey_std, 0)
+
+    return ex_std, ey_std
+
+
+def sse_orig(mx, my, fym, ychat, ymcdat, selwei, model_dat):
+    """weighted MSE target function plus Tikhonov regularization term"""
+
+    # weighted mean squared error
+    ymchat = fym @ ychat
+    err = ymchat - ymcdat
+
+    # sse = torch.trace(err.T @ selwei @ err) # big matrix needs too much RAM
+    # elementwise multiplication and broadcasting and summation:
+    sse = sum(torch.sum(err * err * torch.diag(selwei).view(-1, 1), dim=0))
+
+    # sse with tikhonov term
+    direct = utils.directvec(mx, my, model_dat["idx"], model_dat["idy"])
+    ssetikh = sse + model_dat["alpha"] * direct.T @ direct
+
+    return ssetikh.requires_grad_(True)
